discord_bot/requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.1
supervisor==4.2.5
SQLAlchemy==2.0.29
pydantic==2.7.1
